\name{TentFn}
\alias{TentFn}
\title{
  Tent function dynamics
}
\description{
  This simulates data from a tent function, a simple functionm that generates 'nonlinear' dynamics 
}
\usage{
  TentFn(Nobs, S)
}
\arguments{
  \item{X}{
  Location
}
  \item{S}{
  Tent height (producing non-linear dynamics given S>1.75)
}
}
\details{
}
\value{
  Generates the next observation given location X
}
\references{
}
\author{
  James Thorson
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ ~kwd1 }
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
